Add docs and example
