+++ Examples/StateManagement/redux_tutorial/app/components/CounterList.js
@@ -12,8 +12,7 @@ const CounterList = ({counter, handleAddCounter, handleRemoveCounter, handleIncr
                 handleIncrement={handleIncrement}
                 handleDecrement={handleDecrement}
             />
-
-        ));
+      ));
 
     return (
         <View style={styles.counterFrame}>
