+++ Examples/StateManagement/redux_tutorial/app/reducers/modules/index.js
@@ -1,0 +1,4 @@
+import AddRemoveCounter from './AddRemoveCounter';
+import CountingNumber from './CountingNumber';
+
+export {AddRemoveCounter, CountingNumber};
