+++ Examples/StateManagement/redux_tutorial/index.js
@@ -3,7 +3,7 @@
  */
 
 import {AppRegistry} from 'react-native';
-import EntryPoint from './app/EntryPoint';
+import App from './App';
 import {name as appName} from './app.json';
 
-AppRegistry.registerComponent(appName, () => EntryPoint);
+AppRegistry.registerComponent(appName, () => App);
