+++ docs/state-tutorial/redux-tutorial/01-getting-started/getting-started-kr.md
@@ -1,12 +1,10 @@
----
-layout: default
-title: 1. 시작하기
-parent: Redux 튜토리얼
-grand_parent: State Management Tutorial(한글)
-nav_order: 1
----
 
 ## 1. React Native 를 사용하여 새로운 프로젝트 생성 
+
+```
+react-native init myApp
+cd ya
+```
 
 ### react-native install
