+++ docs/state-tutorial/redux-tutorial/02-Multicounter/kr/01-Let's-make-Multi-Counter-kr.md
@@ -1,0 +1,14 @@
+## 2. Let's make Multi Counter 
+
+ 각각의 카운터에서 숫자를 `increment` , `decrement` 할 수 있고 이 카운터를 여러개 만들 수 있는 `Add`, `Remove` 의 기능이 있는 Multi Counter 예제를 통해 `Redux`를 알아보겠습니다.
+
+### 우리가 만들 것
+.....사진 (완성본)
+
+### Directory 구조
+.... 사진
+
+
+튜토리얼 진행방식은 다음과 같습니다.
+> - 일단 코드를 그대로 따라하신 후, 설명을 읽으면서 이해하기
+> - single Counter 예제 -> Counter 추가, 제거 기능을 통한 Multi Counter 완성
