+++ docs/state-tutorial/redux-tutorial/02-Multicounter/kr/03-Reducer-kr.md
@@ -1,0 +1,105 @@
+## 2-2. 리듀서 `Reducer`
+
+> 이전 state와 action에 따라서, 불변성을 가진 새로운 state를 반환해주는 순수함수.
+
+#### 1. 순수 함수 Reducer 란?
+반환값인 `state` 가 오직 함수 인자인 (state, action)에 의해서 정의되는 함수
+```
+/* es5 */
+function counter(state, action) => {
+	...,
+	return 액션에 따른 state 값;
+}
+
+/* es6+ */
+const counter = (state, action) => {
+	...,
+	return 액션의 따른 state 값;
+}
+```
+
+#### 2. 불변성을 가진 state
+`state` 는 트리 구조이기 때문에 직접 그 값을 변경해주기 위해서 깊이 있는 탐색을 하게 될 수 있고 이에 따라 속성 전체를 비교하면 O(n) 만큼의 데이터를 비교 해야할 수 있습니다. 
+이러한 이유로 `redux` 는 다음과 같은 변경 알고리즘이 적용되어있습니다.
+
+ - reducer를 거친 state의 변경유무를 검사하기 위해 state 객체 주소를 비교하여 다른 주소값이라면 변경됐다고 감지합니다. 
+ - state 값을 변경할 때 직접 값을 수정하지 않고 state의 복제본을 만들어 수정 후, 반환하는 방식을 사용합니다.
+ - 즉, state 값의 변화가 있다면 불변성을 유지해주기 위해서 새로운 객체로 반환해줘야합니다.
+
+```
+리듀서_함수_이름(state, action) {
+	...
+	... (action타입에 따른 state 변화 정의)
+	return state;
+}
+```
+
+#### 해당 프로젝트의 root 위치에서 다음과 같이 파일 생성 후, 수정
+ - 리듀서 함수 정의
+		reducers/index.js
+	```	  
+	/*  
+	* 1. state의 기본 값은 reducer의 인자에서만 선언할 수 있습니다.  
+	* */  
+	const initialState = {  
+	    counter : [  
+	      {  
+	          counterNum : 0,  
+		  },  
+		],  
+	};  
+	  
+	/*  
+	* 2. state는 전체 값을 의미합니다.  
+	*    예를 들어 counterNum에 접근하고 싶다면 `state.counter.counterNum` 으로 접근할 수 있습니다.  
+	* */  
+	  
+	const counter = (state = initialState, action) => {  
+	    const { counter } = state;  
+	  
+		switch(action.type){  
+	        case 'INCREMENT' :  
+	            return({  
+	                counter : [  
+	                    ...counter.slice(0, action.index),  
+						{  
+	                        counterNum : counter[action.index].counterNum + 1,  
+	                    },  
+	                    ...counter.slice(action.index+1, counter.length),  
+	                    ],  
+	            });  
+	  
+	        case 'DECREMENT' :  
+	            return({  
+	                counter : [  
+	                    ...counter.slice(0, action.index),  
+	                    {  
+	                        counterNum : counter[action.index].counterNum - 1,  
+	                    },  
+	                    ...counter.slice(action.index+1, counter.length),  
+	                ],  
+	            });  
+	  
+	        case 'ADD' :  
+	            return({  
+	                counter : [  
+	                    ...counter,  
+	                    {  
+	                       counterNum : 0,  
+	                    },  
+	                ]  
+	            });  
+	  
+	        case 'REMOVE' :  
+	            return({  
+	                counter : counter.slice(0, counter.length - 1),  
+	            });  
+	  
+	        default :  
+	           return state;  
+	  }  
+	}  
+	  
+	export default counter;
+	
+	```
